First cut at GoDaddy connector (not complete). Does a reasonable job of pulling down code, but cannot deploy back to production yet.

+++ src/AlanKent/MagentoDev/Environments/VagrantRsync/VagrantRsyncEnvironment.php
@@ -3,13 +3,14 @@
 namespace AlanKent\MagentoDev\Environments\VagrantRsync;
 
 use AlanKent\MagentoDev\EnvironmentInterface;
+use AlanKent\MagentoDev\MdException;
 
 /**
  * Creates/destroys an environment using 'vagrant rsync-auto' for Magento 2.
  */
 class VagrantRsyncEnvironment implements EnvironmentInterface
 {
-    const SOURCE_DIRNAME = __DIR__.'/VagrantRsync/scripts';
+    const SOURCE_DIRNAME = __DIR__.'/scripts';
 
     const VAGRANTFILE_FILENAME = 'Vagrantfile';
 
@@ -25,21 +26,23 @@ class VagrantRsyncEnvironment implements EnvironmentInterface
      * Create the set of files for the Vagrant based environment.
      * @param array $config
      * @param array $args Currently no command line options are supported.
-     * @return int Exit status code.
+     * @throws MdException Thrown on error.
      */
     public function create(&$config, $args)
     {
         if (count($args) != 0) {
-            echo "Unexpected additional arguments were found.\n";
-            return 1;
+            throw new MdException("Unexpected additional arguments were found.\n", 1);
         }
 
         if (file_exists(self::VAGRANTFILE_FILENAME)) {
-            echo "Please remove the existing '".self::VAGRANTFILE_FILENAME."' before running this command.\n";
-            echo "Or use the 'destroy' command to tear down the current environment.\n";
-            return 1;
+            $msg = "Please remove the existing '".self::VAGRANTFILE_FILENAME."' before running this command.\n";
+            $msg .= "Or use the 'destroy' command to tear down the current environment.\n";
+            throw new MdException($msg, 1);
         }
 
+        if (!file_exists(self::SOURCE_DIRNAME.'/'.self::VAGRANTFILE_FILENAME)) {
+            throw new MdException("Unable to find template Vagrantfile to copy - internal configuration error.\n", 1);
+        }
         file_put_contents(self::VAGRANTFILE_FILENAME, file_get_contents(self::SOURCE_DIRNAME.'/'.self::VAGRANTFILE_FILENAME));
         echo "Created '".self::VAGRANTFILE_FILENAME."'.\n";
 
@@ -47,12 +50,16 @@ class VagrantRsyncEnvironment implements EnvironmentInterface
             mkdir(self::SCRIPTS_DIRNAME);
         }
         if (is_file(self::SCRIPTS_DIRNAME)) {
-            echo "Tried to create a directory called '".self::SCRIPTS_DIRNAME."' but a file with that name already exists.\n";
-            return 1;
+            throw new MdException(
+                "Tried to create a directory called '".self::SCRIPTS_DIRNAME."' but a file with that name already exists.\n",
+                1);
         }
         echo "Created directory '".self::SCRIPTS_DIRNAME."'.\n";
 
         foreach (self::SCRIPTS as $scriptName) {
+            if (!file_exists(self::SOURCE_DIRNAME.'/'.$scriptName)) {
+                throw new MdException("Unable to find script $scriptName to copy - internal configuration error.\n", 1);
+            }
             file_put_contents(self::SCRIPTS_DIRNAME.'/'.$scriptName, file_get_contents(self::SOURCE_DIRNAME.'/'.$scriptName));
             echo "Created '".self::SCRIPTS_DIRNAME.'/'.$scriptName."'.\n";
         }
@@ -64,45 +71,35 @@ class VagrantRsyncEnvironment implements EnvironmentInterface
         echo "    vagrant rsync-auto   Watches for local filesystem changes and copies them\n";
         echo "                         into the VM.\n";
         echo "\n";
-
-        return 0;
     }
 
     /**
      * Destroy the current environment.
      * @param array $config Configuration settings.
      * @param bool $force Set to true if should continue even if something strange occurs.
-     * @return int Exit status code to return.
+     * @throws MdException Thrown on error.
      */
     public function destroy(&$config, $force)
     {
         echo "Running 'vagrant destroy'.\n";
         $exitCode = 0;
-        system('vagrant destroy', $exitCode);
+        system('vagrant destroy -f', $exitCode);
         if (!$force && $exitCode != 0) {
-            return $exitCode;
+            throw new MdException("Failed to shut down Vagrant box.\n", $exitCode);
         }
 
-        $exitCode = self::removeIfUnchanged(self::VAGRANTFILE_FILENAME,
-                                            self::SOURCE_DIRNAME.'/'.self::VAGRANTFILE_FILENAME,
-                                            $force);
-        if ($exitCode != 0) {
-            return $exitCode;
-        }
+        self::removeIfUnchanged(self::VAGRANTFILE_FILENAME,
+                                self::SOURCE_DIRNAME.'/'.self::VAGRANTFILE_FILENAME,
+                                $force);
 
         foreach (self::SCRIPTS as $scriptName) {
-            $exitCode = self::removeIfUnchanged(self::SCRIPTS_DIRNAME.'/'.$scriptName,
-                                                self::SOURCE_DIRNAME.'/'.$scriptName,
-                                                $force);
-            if ($exitCode != 0) {
-                return $exitCode;
-            }
+            self::removeIfUnchanged(self::SCRIPTS_DIRNAME.'/'.$scriptName,
+                                    self::SOURCE_DIRNAME.'/'.$scriptName,
+                                    $force);
         }
 
         echo "Removing directory '".self::SCRIPTS_DIRNAME."'.\n";
         rmdir(self::SCRIPTS_DIRNAME);
-
-        return 0;
     }
 
     /**
@@ -114,23 +111,43 @@ class VagrantRsyncEnvironment implements EnvironmentInterface
      * @param string $localFilename The real filename on disk.
      * @param string $sourceFilename The name of the original file, to compare to the real file.
      * @param bool $force If true, ignore any differences.
-     * @return int The return exit status, where 0 = success.
+     * @throws MdException Thrown on error.
      */
     private static function removeIfUnchanged($localFilename, $sourceFilename, $force)
     {
         if (!file_exists($localFilename)) {
-            return 0;
+            return;
         }
         if (!$force) {
             $localContents = file_get_contents($localFilename);
             $originalContents = file_get_contents($sourceFilename);
             if ($localContents != $originalContents) {
-                echo "Not removing '$localFilename' as it may contain local changes. (Use --force to override.)\n";
-                return 1;
+                throw new MdException("Not removing '$localFilename' as it may contain local changes. (Use --force to override.)\n", 1);
             }
         }
         echo "Removing '$localFilename'.\n";
         unlink($localFilename);
-        return 0;
+    }
+
+    /**
+     * Run the given command inside the environment.
+     * @param array $config Configuration settings to connect to environment.
+     * @param string $cmd The command to run inside the environment.
+     * @throws MdException Thrown on error to report to user.
+     */
+    public function runCommand($config, $cmd)
+    {
+        $run = "sh -c 'vagrant ssh -c \"$cmd\"'";
+        echo "> $run\n";
+        system($run);
+    }
+
+    public function syncToEnvironment($config)
+    {
+        $run = "sh -c 'vagrant rsync'";
+        echo "> $run\n";
+        // If rsync-auto is running in background, give it a chance to complete (hacky!)
+        sleep(3);
+        system($run);
     }
 }
